+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 <!-- markdownlint-disable first-line-h1 -->
+## v1.44.1
+
+*2026-06-10*
+
+**Added**
+
+- [`LanguageService#provideDocumentColors](https://github.com/bhsd-harry/wikiparser-node/wiki/LanguageService-%28EN%29#providedocumentcolors) now supports color parsers that return an object with RGBA components in addition to those that return an array of RGBA values
+
 ## v1.44.0
 
 *2026-06-09*